Move factory methods into their own module
